password-protected ZIP file
The resulting ZIP file requires the provided password to access its contents, providing an additional layer of security. This Python script utilizes the pyzipper library to create a ZIP file that can only be accessed with a password. You can choose the files to include in the ZIP and set a password for encryption. With the provided password, you will be able to access the contents of the ZIP file. This added layer of security ensures that only authorized individuals can access the files within the ZIP.